Try to remove cache only if it is present

+++ fetch_packages.py
@@ -176,11 +176,12 @@ def main(filename):
 
 if __name__ == '__main__':
     os.chdir(os.path.dirname(os.path.realpath(__file__)))
-    try:
-        shutil.rmtree(_PACKAGE_CACHE)
-    except:
-        print "Error: Cannot remove " + _PACKAGE_CACHE
-        exit(1)
+    if os.path.isdir(_PACKAGE_CACHE):
+        try:
+            shutil.rmtree(_PACKAGE_CACHE)
+        except:
+            print "Error: Cannot remove " + _PACKAGE_CACHE
+            exit(1)
 
     try:
         os.makedirs(_PACKAGE_CACHE)
